Разбил объекты на их абстракции и реализацию для группы CPlus
