+++ Abstract_Factory.pro
@@ -7,6 +7,9 @@ CONFIG += c++17 cmdline
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
+        ClassUnit.cpp \
+        MethodUnit.cpp \
+        PrintOperatorUnit.cpp \
         Unit.cpp \
         main.cpp
 
@@ -16,4 +19,7 @@ else: unix:!android: target.path = /opt/$${TARGET}/bin
 !isEmpty(target.path): INSTALLS += target
 
 HEADERS += \
+    ClassUnit.h \
+    MethodUnit.h \
+    PrintOperatorUnit.h \
     Unit.h
