+++ main.cpp
@@ -1,24 +1,26 @@
 #include <QCoreApplication>
-#include "Unit.h"
+#include "ClassUnit.h"
+#include "PrintOperatorUnit.h"
+#include "MethodUnit.h"
 #include <iostream>
 
 
 std::string generateProgram() {
-    ClassUnit myClass( "MyClass" );
+    CPlusClass myClass( "MyClass" );
     myClass.add(
-        std::make_shared< MethodUnit >( "testFunc1", "void", 0 ),
+        std::make_shared< CPlusMethod >( "testFunc1", "void", 0 ),
         ClassUnit::PUBLIC
         );
     myClass.add(
-        std::make_shared< MethodUnit >( "testFunc2", "void", MethodUnit::STATIC ),
+        std::make_shared< CPlusMethod >( "testFunc2", "void", MethodUnit::STATIC ),
         ClassUnit::PRIVATE
         );
     myClass.add(
-        std::make_shared< MethodUnit >( "testFunc3", "void", MethodUnit::VIRTUAL | MethodUnit::CONST ),
+        std::make_shared< CPlusMethod >( "testFunc3", "void", MethodUnit::VIRTUAL | MethodUnit::CONST ),
         ClassUnit::PUBLIC
         );
-    auto method = std::make_shared< MethodUnit >( "testFunc4", "void", MethodUnit::STATIC );
-    method->add( std::make_shared< PrintOperatorUnit >( R"(Hello, world!\n)" ) );
+    auto method = std::make_shared< CPlusMethod >( "testFunc4", "void", MethodUnit::STATIC );
+    method->add( std::make_shared< CPlusPrintOperator >( R"(Hello, world!\n)" ) );
     myClass.add( method, ClassUnit::PROTECTED );
     return myClass.compile();
 }
